Enable multiple motes monitors setup.
Until now, when multiple motes with monitor were added at the same time
they shared the same setup process. As a result the monitor could be
left permanently in a uninitialized state.

This patch bypass the problem by dividing the setup process per mote.
The monitor is considered initialized once the first mote is.

However this is still work in progress. All motes must share the same
platform, that is record/byte offset and endianness.